fix: resolve macOS storage mount/unmount issue
Split CreateRootFS and DestroyVMStorage into platform-specific implementations:

- storage_darwin.go: Creates rootfs image without mount/unmount/initialize steps
  (macOS has no loop mount, hypervisor attaches raw image directly)

- storage_linux.go: Maintains original mount/unmount/initialize behavior

This fixes issue #63 where 'tent create' was hanging on macOS waiting for sudo
password and failing with 'mount -o loop' not found.
